% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntistools-package.R
\docType{package}
\name{ntistools-package}
\alias{ntistools}
\alias{ntistools-package}
\title{ntistools: Tools for Analyzing Nonprofit Trends and Impacts Survey Data}
\description{
Reusable functions for wrangling, analyzing, and visualizing Nonprofit Trends and Impacts Survey (NTIS) data. This package does not contain any raw survey data or PII.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://urbaninstitute.github.io/ntistools/}
  \item \url{https://github.com/UrbanInstitute/ntistools}
  \item Report bugs at \url{https://github.com/UrbanInstitute/ntistools/issues}
}

}
\author{
\strong{Maintainer}: Thiyaghessan Poongundranar \email{tpoongundranar@urban.org}

Other contributors:
\itemize{
  \item Urban Institute [copyright holder]
}

}
\keyword{internal}
