% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_parent.R
\name{propagate_parent}
\alias{propagate_parent}
\title{Push parent variable's special values to child columns}
\usage{
propagate_parent(data, vars, parent_var, values = c(0, 97))
}
\arguments{
\item{data}{A data frame.}

\item{vars}{A character vector of child column names.}

\item{parent_var}{A string giving the name of the parent column.}

\item{values}{Numeric vector of parent values to propagate to children.
Defaults to \code{c(0, 97)}.}
}
\value{
\code{data} with child columns updated.
}
\description{
When a parent variable has one of the specified values (e.g., 0 or 97), set
that same value in all child columns. This is useful when a parent question
gates a set of follow-up questions.
}
\examples{
d <- data.frame(parent = c(0, 1, 97, 1), child1 = c(NA, 1, NA, 0))
propagate_parent(d, "child1", parent_var = "parent", values = c(0, 97))
}
